Fix Buchungsart Steuer Long statt Int (#449)

+++ src/de/jost_net/JVerein/io/KontenrahmenImportXML.java
@@ -128,7 +128,7 @@ public class KontenrahmenImportXML implements Importer
             .createList(Buchungsart.class);
         stbuait.addFilter("nummer = ?", mapst.get(id));
         Buchungsart stbua = stbuait.next();
-        bua.setSteuerBuchungsart(Integer.parseInt(stbua.getID()));
+        bua.setSteuerBuchungsart(Long.parseLong(stbua.getID()));
         bua.store();
       }
     }
